requirements.txt
coverage==3.6
flufl.enum==3.3.2
nose==1.2.1
unittest2==0.5.1
wsgiref==0.1.2
tornado==3.1.1
sockjs-tornado==1.0.0
